+++ app/src/main/java/com/example/ProvasTabletActivity.java
@@ -1,0 +1,16 @@
+package com.example;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.listaalunos.R;
+
+public class ProvasTabletActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_provas_tablet);
+    }
+}
