+++ app/src/main/res/layout/activity_provas.xml
@@ -1,14 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/frame_principal"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:cardElevation="5dp"
-    app:cardUseCompatPadding="true">
+    android:layout_height="match_parent">
 
-    <ListView
-        android:id="@+id/provas_lista"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
 
-</androidx.cardview.widget.CardView>
+</FrameLayout>
